+++ packages/client/src/index.ts
@@ -22,6 +22,7 @@ export type {
   ApplyBatchActionOptions,
   NOOP,
   OsdkActionParameters,
+  WhereClause,
 } from "@osdk/client.api";
 export { PalantirApiError } from "@osdk/shared.net.errors";
 
@@ -36,6 +37,5 @@ export type { InterfaceObjectSet, ObjectSet } from "./objectSet/ObjectSet.js";
 export type { OsdkObject } from "./OsdkObject.js";
 export type { Osdk } from "./OsdkObjectFrom.js";
 export type { PageResult } from "./PageResult.js";
-export type { WhereClause } from "./query/WhereClause.js";
 export { isOk } from "./ResultOrError.js";
 export type { ResultOrError } from "./ResultOrError.js";
