+++ packages/client/src/internal/conversions/legacyToModernSingleAggregationResult.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
+import type { AggregationClause } from "@osdk/client.api";
 import type { AggregateObjectsResponseV2 } from "@osdk/internal.foundry";
 import invariant from "tiny-invariant";
 import type { AggregationResultsWithoutGroups } from "../../query/aggregations/AggregationResultsWithoutGroups.js";
-import type { AggregationClause } from "../../query/aggregations/AggregationsClause.js";
 import type { ArrayElement } from "../../util/ArrayElement.js";
 
 export function legacyToModernSingleAggregationResult<
