+++ packages/client/src/objectSet/ObjectSet.ts
@@ -21,10 +21,12 @@ import type {
   ObjectTypeDefinition,
 } from "@osdk/api";
 import type {
+  AggregateOpts,
+  AggregateOptsThatErrors,
   BaseObjectSet,
   PropertyValueClientToWire,
+  WhereClause,
 } from "@osdk/client.api";
-import type { AggregateOptsThatErrors } from "../object/AggregateOptsThatErrors.js";
 import type {
   Augments,
   FetchPageArgs,
@@ -38,9 +40,7 @@ import type {
 } from "../object/FetchPageResult.js";
 import type { Result } from "../object/Result.js";
 import type { Osdk } from "../OsdkObjectFrom.js";
-import type { AggregateOpts } from "../query/aggregations/AggregateOpts.js";
 import type { AggregationsResults } from "../query/aggregations/AggregationsResults.js";
-import type { WhereClause } from "../query/WhereClause.js";
 import type { LinkedType, LinkNames } from "./LinkUtils.js";
 
 export interface MinimalObjectSet<Q extends ObjectOrInterfaceDefinition>
